Run CellChat with OmniPath function [[DEPRECATED]]

Run CellChat with OmniPath function [[DEPRECATED]]

call_cellchat(
  sce,
  op_resource,
  .format = TRUE,
  exclude_anns = c(),
  nboot = 100,
  assay = "RNA",
  .seed = 1004,
  .normalize = FALSE,
  .do_parallel = FALSE,
  .raw_use = TRUE,
  expr_prop = 0,
  organism = "human",
  thresh = 1,
  de_thresh = 0.05,
  ...
)

Arguments

sce

Seurat object as input

op_resource

OmniPath Intercell Resource DN

.format

bool whether to format output

exclude_anns

Annotation criteria to be excluded

nboot

number of bootstraps to calculate p-value

assay

assay name (RNA by default)

.seed

random seed

.normalize

# bool whether to normalize non-normalized data with

.do_parallel

whether to parallelize or not

.raw_use

whether use the raw data or gene expression data projectected to a ppi (should be kept to TRUE)

expr_prop

minimum proportion of gene expression per cell type (0 by default), yet perhaps one should consider setting this to an appropriate value between 0 and 1, as an assumptions of these method is that communication is coordinated at the cluster level.

organism

Obtain CellChatDB for which organism ('mouse' or 'human')

thresh

p-value threshold (1 by default)

de_thresh

diff expression of genes p-value

...

Arguments passed on to CellChat::subsetCommunication

object

CellChat object

net

Alternative input is a data frame with at least with three columns defining the cell-cell communication network ("source","target","interaction_name")

slot.name

the slot name of object: slot.name = "net" when extracting the inferred communications at the level of ligands/receptors; slot.name = "netP" when extracting the inferred communications at the level of signaling pathways

sources.use

a vector giving the index or the name of source cell groups

targets.use

a vector giving the index or the name of target cell groups.

signaling

a character vector giving the name of signaling pathways of interest

pairLR.use

a data frame consisting of one column named either "interaction_name" or "pathway_name", defining the interactions of interest

datasets

select the inferred cell-cell communications from a particular `datasets` when inputing a data frame `net`

ligand.pvalues,ligand.logFC,ligand.pct.1,ligand.pct.2

set threshold for ligand genes

ligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)

ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genes

ligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1

ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group

receptor.pvalues,receptor.logFC,receptor.pct.1,receptor.pct.2

set threshold for receptor genes

Value

A DF of intercellular communication network

Details

CellChat's objects are not lazily documented/exported thus the whole package has to be imported.