Call Squidpy Pipeline via reticulate with OmniPath and format results

Call Squidpy Pipeline via reticulate with OmniPath and format results

call_squidpy(
  seurat_object,
  op_resource,
  seed = 1004,
  conda_env = NULL,
  assay = "RNA",
  assay.type = "logcounts",
  ...
)

Arguments

seurat_object
Seurat object as input

op_resource
Tibble or list of OmniPath resources, typically obtained via select_resource

seed
seed passed to squidpy's ligrec function

conda_env
python conda environment to run Squidpy; set to liana_env by default

assay
assay name

...
kwargs passed to Squidpy; For more information see: https://squidpy.readthedocs.io/en/latest/api/squidpy.gr.ligrec.html#squidpy.gr.ligrec

expr_prop
minimum proportion of gene expression per cell type (0 by default), yet perhaps one should consider setting this to an appropriate value between 0 and 1, as an assumptions of these method is that communication is coordinated at the cluster level.

Value

A list of Squidpy results for each resource

Details

CellPhoneDBv2 algorithm re-implementation in Python.

Note that `cluster_key` is a parameter passed to the Squidpy function, by default this will be set to the default Ident of the Seurat object.