Function to get unfiltered intercell resources For each resource and OmniPath variant compiles tables of ligands, receptors and interactions

Function to get unfiltered intercell resources For each resource and OmniPath variant compiles tables of ligands, receptors and interactions

compile_ligrec(lr_pipeline = TRUE)

Arguments

lr_pipeline
bool whether to format for lr_pipeline and remove

duplicate LRs (mainly from composite OmniDB due to category (adhesion vs lr))

omni_variants
bool whether to get different OmniPath variants (e.g. _full, based on ligrec resource quality quartile, or if only lig_rec)

Value

A list of OmniPath resources formatted according to the method pipes

Details

calls on omnipath_intercell, intercell_connections, get_partners, and intercell_connections