Function to convert a list of characters to a ranked matrix [0,1]

Function to convert a list of characters to a ranked matrix [0,1]

.rank_matrix(glist, ...)

Arguments

glist

a list of ranked/ordered characters

Value

a matrix filled with 0-1 values depending on the position order of the characters.

Details

Adated from Kolde et al., 2012. Required due to the removal of the RobustRankAggregate package from CRAN.

References

Kolde, R., Laur, S., Adler, P. and Vilo, J., 2012. Robust rank aggregation for gene list integration and meta-analysis. Bioinformatics, 28(4), pp.573-580.