Communication Frequency heatmap plot

Communication Frequency heatmap plot

heat_freq(liana_res, ...)

Arguments

liana_res
aggregated liana results (preferably truncated to some threshold)

...
Arguments passed on to liana_heatmap

mat

Diagonal celltype-celltype matrix to be plotted. In theory, any metric deemed meaningful between cell pairs can be plotted.

font_size

base font_size - other fontsizes are relative to this one

grid_text

logical whether to display grid text or not

name

name of the heatmap. By default the heatmap name is used as the title of the heatmap legend.

row_title

Row tittle

column_title

Column tittle

Details

This plot was inspired by CellPhoneDB and also CellChat's heatmap design. It makes the assumption that the number of interactions inferred between cell types is informative of the communication events occurring in the system as a whole. This is a rather strong assumption limited by the arbitrarily filtered interactions Thus, I suggest that one limits any conclusions, unless supported by complimentary information, such as biological prior knowledge.