Retrieves the interactions from one ligand-receptor resource

Retrieves the interactions from one ligand-receptor resource

intercell_connections(resource, ...)

Arguments

resource
Name of current resource (taken from get_lr_resources)

...
Arguments passed on to OmnipathR::import_post_translational_interactions

resources

interactions not reported in these databases are removed. See get_interaction_resources for more information.

organism

Interactions are available for human, mouse and rat. Choose among: 9606 human (default), 10116 rat and 10090 Mouse

exclude

Character: datasets or resources to exclude.

references_by_resource

if FALSE, removes the resource name prefixes from the references (PubMed IDs); this way the information which reference comes from which resource will be lost and the PubMed IDs will be unique.