Liana dotplot interactions by source and target cells

Liana dotplot interactions by source and target cells

liana_dotplot(
  liana_agg,
  source_groups,
  target_groups,
  specificity = "natmi.edge_specificity",
  magnitude = "sca.LRscore",
  y.label = "Interactions (Ligand -> Receptor)",
  size.label = "Interaction\nSpecificity",
  colour.label = "Expression\nMagnitude",
  show_complex = TRUE
)

Arguments

liana_agg
aggregated `liana_wrap` results -> preferentially filtered by some condition (e.g. preferential ranking, specific interactions, etc)

source_groups
names of the source(sender) cell types

target_groups
names of the target cell types

specificity
column to represent the dot-size of the interaction (by default `natmi.edge_specificity`)

magnitude
column to represent interactions expression magnitude (by default `sca.LRscore`)

y.label
y label name

size.label
size (~specificty) label name

colour.label
colour (~magnitude) label name

show_complex
logical whether to show complexes (default - TRUE) or only the subunit with minimum expression.

Value

a ggplot2 object

Details

Here, we refer to `specificity` as how specific this interaction is to a cell type pair regards to the rest of the cell type pairs ( e.g. CellPhoneDB's p-values, NATMI's specificity edges, Connectome's scaled weights, etc)

`magnitude` on the other hand is a direct measure of the expression alone, by default we use SingleCellSignalR's dataset indepent LRscore (bound between 0 and 1). Yet, one could also use CellChat's probabilities or CellPhoneDB's means, etc.