Liana Pipe which runs DE analysis and merges needed information for LR inference

Liana Pipe which runs DE analysis and merges needed information for LR inference

liana_pipe(
  sce,
  op_resource,
  test.type = "wilcox",
  pval.type = "all",
  assay = "RNA",
  assay.type = "logcounts"
)

Arguments

sce
SingleCellExperiment Object

op_resource
resource tibble obtained via liana::select_resource

test.type
String specifying the type of pairwise test to perform - a t-test with "t", a Wilcoxon rank sum test with "wilcox", or a binomial test with "binom".

pval.type
A string specifying how p-values are to be combined across pairwise comparisons for a given group/cluster.

assay
assay to be used ("RNA" by default)

assay.type
- the type of data to be used to calculate the means (logcounts by default), available options are: "counts" and "logcounts"

expr_prop
minimum proportion of gene expression per cell type (0.2 by default), yet one should consider setting this to an appropriate value between 0 and 1, as an assumptions of these method is that communication is coordinated at the cluster level.

Value

Returns a tibble with information required for LR calculations downstream