Run LIANA Wrapper

Run LIANA Wrapper

liana_wrap(
  seurat_object,
  method = c("natmi", "connectome", "logfc", "cellchat", "sca", "squidpy"),
  resource = c("OmniPath"),
  external_resource,
  .simplify = TRUE,
  ...
)

Arguments

seurat_object
seurat object

method
method(s) to be run via liana

resource
resource(s) to be used by the methods (Use `all` to run all resources in one go), or `custom` to run liana_wrap with an appropriately formatted custom resource, passed via `exernal_resource`

external_resource
external resource in OmniPath tibble format

.simplify
if methods are run with only 1 resource, return a list of tibbles for each method (default), rather than a list of lists with method-resource combinations

...
Arguments passed on to liana_defaults

cellchat.params

list of Parameters passed to CellChat call_cellchat

sca.params

list of Parameters passed to SingleCellSignalR call_sca

squidpy.params

list of Parameters passed to Squidpy call_squidpy

call_connectome.params

list of Parameters passed to Connectome call_connectome

call_italk.params

list of Parameters passed to iTALK call_italk

call_natmi.params

list of Parameters passed to NATMI call_natmi

liana_pipe.params

list of Parameters passed to NATMI liana_pipe

liana_call.params

list of Parameters passed to NATMI liana_call

assay

Assay name passed to `call_italk`, `call_sca`, `call_cellchat`, and `call_connectome`

decomplexify

specify whether complexes in the resource should be dissociated and taken into account

expr_prop

minimum proportion of gene expression per cell type (0 by default), yet perhaps one should consider setting this to an appropriate value between 0 and 1, as an assumptions of these method is that communication is coordinated at the cluster level.

Value

A list of method-resource results - i.e. provided resources are run with each method If only one resource is selected, a single tibble (with results for that resource) will be returned for each of the selected methods

Details

LIANA wrapper method that can be used to call each method with a given set of intercellular resources from the OmniPath universe.