CellChat x Omni Pipe function

CellChat x Omni Pipe function

call_cellchat(
  op_resource,
  seurat_object,
  .format = TRUE,
  exclude_anns = c(),
  nboot = 100,
  assay = "RNA",
  .seed = 1004,
  .normalize = FALSE,
  .do_parallel = FALSE,
  .raw_use = TRUE,
  ...
)

Arguments

op_resource
OmniPath Intercell Resource DN

seurat_object
Seurat object as input

.format
bool whether to format output

exclude_anns
Annotation criteria to be excluded

nboot
number of bootstraps to calculate p-value

.normalize
# bool whether to normalize non-normalized data with

.raw_use
whether use the raw data or gene expression data pojectected to a ppi

...
Arguments passed on to CellChat::subsetCommunication

object

CellChat object

net

Alternative input is a data frame with at least with three columns defining the cell-cell communication network ("source","target","interaction_name")

slot.name

the slot name of object: slot.name = "net" when extracting the inferred communications at the level of ligands/receptors; slot.name = "netP" when extracting the inferred communications at the level of signaling pathways

sources.use

a vector giving the index or the name of source cell groups

targets.use

a vector giving the index or the name of target cell groups.

signaling

a character vector giving the name of signaling pathways of interest

pairLR.use

a data frame consisting of one column named either "interaction_name" or "pathway_name", defining the interactions of interest

thresh

threshold of the p-value for determining significant interaction

datasets

select the inferred cell-cell communications from a particular `datasets` when inputing a data frame `net`

ligand.pvalues

set threshold for ligand genesligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genesligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group

ligand.logFC

set threshold for ligand genesligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genesligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group

ligand.pct.1

set threshold for ligand genesligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genesligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group

ligand.pct.2

set threshold for ligand genesligand.pvalues: threshold for pvalues in the differential expression gene analysis (DEG)ligand.logFC: threshold for logFoldChange in the DEG analysis; When ligand.logFC > 0, keep upgulated genes; otherwise, kepp downregulated genesligand.pct.1: threshold for the percent of expressed genes in the defined 'positive' cell group. keep genes with percent greater than ligand.pct.1ligand.pct.2: threshold for the percent of expressed genes in the cells except for the defined 'positive' cell group

receptor.pvalues

set threshold for receptor genes

receptor.logFC

set threshold for receptor genes

receptor.pct.1

set threshold for receptor genes

receptor.pct.2

set threshold for receptor genes

Value

A DF of intercellular communication network

Details

CellChat's objects are not documented/exported thus the whole package has to be imported.