Function to call connectome with databases from OmniPath

Function to call connectome with databases from OmniPath

call_connectome(
  seurat_object,
  op_resource,
  .format = TRUE,
  .spatial = FALSE,
  ...
)

Arguments

seurat_object
Seurat object as input

op_resource
OmniPath Intercell Resource DN

.format
bool whether to format output

...
Arguments passed on to Connectome::CreateConnectome

object

A Seurat object

LR.database

Accepts either 'fantom5' or 'custom'. If custom, a dataframe must be provided to argument custom.list with the first column equal to ligands, second column equal to associated receptors, and third column equal to desired modal categorizations.

species

The species of the object that is being processed. Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'

include.putative

Default TRUE. Includes ligand-receptor pairs deemed putative in FANTOM5 database.

include.rejected

Default FALSE. If TRUE, includes gene pairs labeled "EXCLUDED" in FANTOM5 database. See ncomms8866 .rda file for qualifications for exclusion.

p.values

Default TRUE. Runs a Wilcoxon Rank test to calculate adjusted p-value for ligand and receptor expression within the input object. Change to FALSE for decreased run-time.

max.cells.per.ident

Default NULL. If a value is input, input object will be downsampled to requested number of cells per identity. This can greatly improve run-time.

min.cells.per.ident

Default NULL. If a value is input, only cell populations meeting this threshold will be included in network analysis. Can limit high-variation effects from small clusters.

weight.definition.norm

Method of edgeweight definition from normalized data slot. Either 'sum','mean',or 'product'. Defaults to 'product'. 'Sum' adds values from sending and receiving clusters, 'mean' averages them, and 'product' multiplies them.

weight.definition.scale

Method of edgeweight definition from scaled data slot. Either 'sum','mean',or 'product'. Defaults to 'mean'. 'Sum' adds values from sending and receiving clusters, 'mean' averages them, and 'product' multiplies them.

custom.list

Optional. A dataframe for custom mapping, with the first column equal to ligands, second column equal to associated receptors, and third column equal to desired modal categorizations. If modal categorizations are unknown, fill with 'UNCAT' or similar placeholder.

calculate.DOR

Whether to include log normalized Diagnostic Odds Ratio in the output. Default FALSE.

assay

Which assay to employ to create connectomic mapping. Default = 'RNA', also accepts 'SCT' and others

Value

An unfiltered connectome results df

Details

Stats: 1) The ‘weight_norm’ edge attribute is derived from the normalized expression of the ligand and the receptor in the single-cell data. 2) The ‘weight_scale’ edge attribute is derived from the z-scores of the ligand and the receptor in each edge, and is of higher value when the ligand and receptor are more specific to a given pair of cell types 3) DEG p-values for L and R