Robust Aggregate ranks using `RobustRankAggreg`

Robust Aggregate ranks using `RobustRankAggreg`

.aggregate_rank(liana_mlist, ...)

Arguments

liana_mlist
liana list with method tibbles

...
Arguments passed on to RobustRankAggreg::aggregateRanks

glist

list of element vectors, the order of the vectors is used as the ranking.

rmat

the rankings in matrix format. The glist is by default converted to this format.

N

the number of ranked elements, important when using only top-k ranks, by default it is calculated as the number of unique elements in the input.

method

rank aggregation method, by defaylt 'RRA', other options are 'min', 'geom.mean', 'mean', 'median' and 'stuart'

full

indicates if the full rankings are given, used if the the sets of ranked elements do not match perfectly

exact

indicator showing if exact p-value will be calculated based on rho score (Default: if number of lists smaller than 10, exact is used)

topCutoff

a vector of cutoff values used to limit the number of elements in the input lists elements do not match perfectly