Helper Function to do prop filtering of liana_pipe output

Helper Function to do prop filtering of liana_pipe output

.filt_liana_pipe(liana_res, method, ...)

Arguments

liana_res

resutls from `liana_pipe`

method

current method

...

Arguments passed on to liana_defaults

liana_pipe.params

list of Parameters passed to NATMI liana_pipe

liana_call.params

list of Parameters passed to NATMI liana_call

cellchat.params

list of Parameters passed to CellChat call_cellchat

squidpy.params

list of Parameters passed to Squidpy call_squidpy

call_connectome.params

list of Parameters passed to Connectome call_connectome

call_italk.params

list of Parameters passed to iTALK call_italk

call_natmi.params

list of Parameters passed to NATMI call_natmi

call_sca.params

list of Parameters passed to SingleCellSignalR call_sca

assay

Assay name passed to `call_italk`, `call_sca`, `call_cellchat`, and `call_connectome`

cellphonedb.params

list of Parameters passed to liana's internal cellphonedb implementation cellphonedb_score

natmi.params

list of Parameters passed to liana's internal edge_specificity implementation natmi_score

sca.params

list of Parameters passed to liana's internal LRScore implementation sca_score

connectome.params

list of Parameters passed to liana's internal connectome's weight_sc implementation connectome_score

cytotalk.params

list of Parameters passed to liana's internal crosstalk scores implementation cytotalk_score

logfc.params

list of Parameters passed to liana's internal logFC implementation logfc_score

permutation.params

list of parameters passed to permutation methods

expr_prop

minimum proportion of gene expression per cell type (0.1 by default). This is further applied at a method-specific level - it is not applied to Connectome and Cytotalk, as these methods calculate/normalize the score (or statistics used in their scores) across all cell types or cell type pairs.

complex_policy

policy how to account for the presence of complexes.

seed

random seed integer

parallelize

whether to parallelize cellphonedb-like

workers

number of workers to be called

assay.type

- the type of data to be used to calculate the means (logcounts by default), available options are: "counts" and "logcounts"