Function to Aggregate CCC Method Results

Function to Aggregate CCC Method Results

liana_aggregate(
  liana_res,
  resource = NULL,
  set_cap = "max",
  cap = NULL,
  get_ranks = TRUE,
  get_agrank = TRUE,
  .score_mode = .score_specs,
  ...
)

Arguments

liana_res
LIANA results

resource
If methods are ran with multiple resources, the name of the resource of interest needs to be provided *Note* if a name is not provided, the first results based on the first resource in the list will be returned

set_cap
Function used to set ranked cap (i.e. the value that is assigned to interactions with NA for scores); By default, this is set to "max", which is the maximum number of interactions obtained by the methods; Some methods return all possible ligand-receptor combinations for each possible source and target cell pair - i.e. the known universe of all possible interactions (based on the CCC resource)

cap
A cap can for all methods can also be manually set, then the top X interactions, based on the `specificity` scores for each method will be returned and the ranking will be carried out solely on them

get_ranks
boolean, whether to return consensus ranks for methods

get_agrank
boolean, whether to return aggregate rank using the `RobustRankAggreg` package.

.score_mode
defines the way that the methods would be aggragate. By default, we use the score of each method which reflects specificity (if available), if not e.g. the case of SCA we use it's sole scoring function. This aggregation is by default done on the basis of the list returns by `.score_mode`. Alternatively, one could pass `.score_housekeep` to obtain an aggragate of the housekeeping interactions of each `external` LIANA++ method.

...
Arguments passed on to RobustRankAggreg::aggregateRanks

glist

list of element vectors, the order of the vectors is used as the ranking.

rmat

the rankings in matrix format. The glist is by default converted to this format.

N

the number of ranked elements, important when using only top-k ranks, by default it is calculated as the number of unique elements in the input.

method

rank aggregation method, by defaylt 'RRA', other options are 'min', 'geom.mean', 'mean', 'median' and 'stuart'

full

indicates if the full rankings are given, used if the the sets of ranked elements do not match perfectly

exact

indicator showing if exact p-value will be calculated based on rho score (Default: if number of lists smaller than 10, exact is used)

topCutoff

a vector of cutoff values used to limit the number of elements in the input lists elements do not match perfectly

Value

Tibble with the interaction results and ranking for each method

Details

set_cap is the name of the name of a function that is to be executed on a vector representing the number of rows in the results for each method, by default this is set to base::max, but any other function that works with vectors could be passed - e.g. min, mean, etc.

This function also decomplexifies any complex present in the CellChat results which returns complexes by default