Function to pass Default Arguments for each method

Function to pass Default Arguments for each method

Usage

liana_defaults(
  assay = "RNA",
  assay.type = "logcounts",
  expr_prop = 0.1,
  seed = 1004,
  complex_policy = "mean0",
  parallelize = FALSE,
  workers = 8,
  permutation.params = NULL,
  liana_pipe.params = NULL,
  liana_call.params = NULL,
  cellphonedb.params = NULL,
  natmi.params = NULL,
  sca.params = NULL,
  connectome.params = NULL,
  cytotalk.params = NULL,
  logfc.params = NULL,
  cellchat.params = NULL,
  squidpy.params = NULL,
  call_sca.params = NULL,
  call_natmi.params = NULL,
  call_connectome.params = NULL,
  call_italk.params = NULL
)

Arguments

assay

Assay name passed to `call_italk`, `call_sca`, `call_cellchat`, and `call_connectome`

assay.type

- the type of data to be used to calculate the means (logcounts by default), available options are: "counts" and "logcounts"

expr_prop

minimum proportion of gene expression per cell type (0.1 by default). Note that when working with complexes, the minimum subunit proportion will be used for filtering.

seed

random seed integer

complex_policy

policy how to account for the presence of complexes.

parallelize

whether to parallelize cellphonedb-like

workers

number of workers to be called

permutation.params

list of parameters passed to permutation methods

liana_pipe.params

list of Parameters passed to NATMI liana_pipe

liana_call.params

list of Parameters passed to NATMI liana_call

cellphonedb.params

list of Parameters passed to liana's internal cellphonedb implementation cellphonedb_score

natmi.params

list of Parameters passed to liana's internal edge_specificity implementation natmi_score

sca.params

list of Parameters passed to liana's internal LRScore implementation sca_score

connectome.params

list of Parameters passed to liana's internal connectome's weight_sc implementation connectome_score

cytotalk.params

list of Parameters passed to liana's internal crosstalk scores implementation cytotalk_score

logfc.params

list of Parameters passed to liana's internal logFC implementation logfc_score

cellchat.params

list of Parameters passed to CellChat call_cellchat

squidpy.params

list of Parameters passed to Squidpy call_squidpy

call_sca.params

list of Parameters passed to SingleCellSignalR call_sca

call_natmi.params

list of Parameters passed to NATMI call_natmi

call_connectome.params

list of Parameters passed to Connectome call_connectome

call_italk.params

list of Parameters passed to iTALK call_italk

Value

A list of the default parameters for each method

Details

The default parameters for each method can also be overwritten by manually passing a list of parameters for the appropraite method liana_wrap

Further, each `get_*` method will by default obtain the default params passed via liana_pipe and liana_call. This is done so that most steps required for the calculation of these methods are undertaken only once.

NB! LIANA's internal methods are made consistent. There is no reason to pass specific parameters to any of them. Thus, it is best that one sticks to the non-nested parameters of this function (i.e. excluding `.params`), unless a very specific reason requires any of LIANA's internal parameters to be changed.

Examples

liana_path <- system.file(package = "liana")
# load testdata
testdata <- readRDS(file.path(liana_path , "testdata", "input", "testdata.rds"))
# get a `named` list with all default parameters passed to liana.
def_params <- liana_defaults()
# any of these can then be overwritten and passed to `...` in `liana_wrap`
# with the `.params` suffix to the parameter name type. For example,
liana_res <- liana_wrap(testdata,
                        permutation.params = list(nperms=2),
                        liana_pipe.params = list(test.type='wilcox'))
#> Expression from the `RNA` assay will be used
#> Running LIANA with `seurat_annotations` as labels!
#> Warning: 3465 genes and/or 0 cells were removed as they had no counts!
#> LIANA: LR summary stats calculated!
#> Now Running: Natmi
#> Now Running: Connectome
#> Now Running: Logfc
#> Now Running: Sca
#> Now Running: Cellphonedb